+++ CHANGELOG.md
@@ -19,6 +19,7 @@ New:
   * of `C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5` (Windows Application Packaging)
 * #60 - Show error when `SolutionGuid` is used by a project
 * #60 - Show error when `SolutionGuid` is reversed by a project type
+* #64 - Show error for missing parameters in project lines
 
 Improvement: